+++ XFaction/Core/DataText/DTGuild.lua
@@ -167,8 +167,23 @@ local function LineClick(_, inUnitGUID, inMouseButton)
 			-- Who
 			SetItemRef('player:' .. _UnitName, format('|Hplayer:%1$s|h[%1$s]|h', _UnitName), 'LeftButton') 
 		elseif(_UnitFaction:Equals(_PlayerFaction)) then
-			-- Whisper		
-			SetItemRef('player:' .. _UnitName, format('|Hplayer:%1$s|h[%1$s]|h', _UnitName), 'LeftButton')
+			-- Whisper
+			if(XFG.Player.Faction:Equals(_Unit:GetFaction())) then
+				SetItemRef('player:' .. _UnitName, format('|Hplayer:%1$s|h[%1$s]|h', _UnitName), 'LeftButton')
+			-- See if theyre a bnet friend
+			else
+				for _, _Friend in XFG.Friends:Iterator() do
+					if(_Unit:GetName() == _Friend:GetName()) then
+						local _Target = _Friend:GetTarget()
+						local _Realm = _Target:GetRealm()
+						if(_Realm:Equals(_Unit:GetRealm())) then
+							local _Name = _Friend:GetAccountName()
+							--"|HBNplayer:"..arg2..":"..arg13..":"..arg11..":"..chatGroup..(chatTarget and ":"..chatTarget or "").."|h"
+							SetItemRef('player:'.. _UnitName, format('|HBNplayer:%1:%2:0:1$s|h[%1$s]|h', _Friend:GetAccountName(), _Friend:GetID()), 'LeftButton')
+						end
+					end
+				end
+			end
 		end		
 	elseif inMouseButton == 'RightButton' then
 		if IsShiftKeyDown() then