+++ XFaction/Core/Handlers/ChannelEvent.lua
@@ -23,10 +23,6 @@ function ChannelEvent:Initialize()
         XFG:Info(LogCategory, "Registered for CHAT_MSG_CHANNEL_NOTICE events")
 		XFG:RegisterEvent('CHANNEL_FLAGS_UPDATED', self.CallbackChannelChange)
         XFG:Info(LogCategory, "Registered for CHANNEL_FLAGS_UPDATED events")
-		XFG:RegisterEvent('CHAT_SERVER_DISCONNECTED', self.CallbackDisconnect)
-		XFG:Info(LogCategory, "Registered for CHAT_SERVER_DISCONNECTED events")
-		XFG:RegisterEvent('CHAT_SERVER_RECONNECTED', self.CallbackReconnect)
-		XFG:Info(LogCategory, "Registered for CHAT_SERVER_RECONNECTED events")	
 		XFG:RegisterEvent('CHAT_MSG_CHANNEL_LEAVE', self.CallbackOffline)
 		XFG:Info(LogCategory, "Registered for CHAT_MSG_CHANNEL_LEAVE events")
 		self:IsInitialized(true)
@@ -60,31 +56,30 @@ function ChannelEvent:SetKey(inKey)
 end
 
 function ChannelEvent:CallbackChannelNotice(inAction, _, _, inChannelName, _, _, inChannelType, inChannelNumber, inChannelShortName)
+	local _Channel = XFG.Outbox:GetLocalChannel()
 	-- Fires when player leaves a channel
-	if(inAction == 'YOU_LEFT') then
-		if(XFG.Channels:RemoveChannel(inChannelShortName)) then
-			local _Channel = XFG.Outbox:GetLocalChannel()
-			if(_Channel:GetShortName() == inChannelShortName) then
-				XFG:Error(LogCategory, "Removed channel was the addon channel")
-			end
-		end
+	if(inAction == 'YOU_LEFT' and XFG.Settings.Network.Channel.Name == inChannelShortName) then
+		XFG.Channels:RemoveChannel(inChannelShortName)
+		XFG:Error(LogCategory, "Removed channel was the addon channel")
 
 	-- Fires when player joins a channel
-	elseif(inAction == 'YOU_CHANGED') then
-		XFG.Channels:ScanChannels()
+	elseif(inAction == 'YOU_CHANGED' and XFG.Settings.Network.Channel.Name == inChannelShortName) then
+		if(XFG.Outbox:HasLocalChannel()) then
+			local _Channel = XFG.Outbox:GetLocalChannel()
+			_Channel:SetID(inChannelNumber)
+		else
+			local _NewChannel = Channel:new()
+			_NewChannel:SetKey(inChannelShortName)
+			_NewChannel:SetID(inChannelNumber)
+			_NewChannel:SetShortName(inChannelShortName)
+			XFG.Channels:AddChannel(_NewChannel)
+			XFG.Outbox:SetLocalChannel(_NewChannel)
+		end
 	end
 end
 
 function ChannelEvent:CallbackChannelChange(inIndex)
 	XFG.Channels:ScanChannel(inIndex)
-end
-
-function ChannelEvent:CallbackDisconnect()
-	XFG:Info(LogCategory, "Received CHAT_SERVER_DISCONNECTED system event")
-end
-
-function ChannelEvent:CallbackReconnect()
-	XFG:Info(LogCategory, "Received CHAT_SERVER_RECONNECTED system event")
 end
 
 function ChannelEvent:CallbackOffline(_, inUnitName, _, _, _, _, _, inChannelID, _, _, _, inGUID)
@@ -95,8 +90,6 @@ function ChannelEvent:CallbackOffline(_, inUnitName, _, _, _, _, _, inChannelID,
 		local _UnitData = XFG.Confederate:GetUnit(inGUID)
 		XFG.Confederate:RemoveUnit(inGUID)
 		XFG.Frames.System:DisplayLocalOffline(_UnitData)
-		XFG.Links:RemoveNode(_UnitData:GetName())
-		XFG.DataText.Guild:RefreshBroker()
-		XFG.DataText.Links:RefreshBroker()
+		XFG.Links:RemoveNode(_UnitData:GetName())		
 	end
 end