+++ XFaction/Initialize.lua
@@ -9,37 +9,6 @@ function XFG:Init()
 	XFG.Guilds = GuildCollection:new(); XFG.Guilds:Initialize()
 	XFG.Teams = TeamCollection:new(); XFG.Teams:Initialize()
 
-	-- -- Make sure we have all the realm/guild combinations accounted for
-	-- for _RealmName, _FactionGuilds in pairs(XFG.Settings.Guilds) do
-	-- 	XFG:Debug(LogCategory, 'Initializing realm [%s]', _RealmName)
-	-- 	local _NewRealm = Realm:new()
-	-- 	_NewRealm:SetKey(_RealmName)
-	-- 	_NewRealm:SetName(_RealmName)
-	-- 	_NewRealm:SetAPIName(string.gsub(_RealmName, '%s+', ''))
-	-- 	_NewRealm:Initialize()
-	-- 	XFG.Realms:AddRealm(_NewRealm)
-	-- 	for _FactionName, _Guilds in pairs(_FactionGuilds) do
-	-- 		local _Faction = XFG.Factions:GetFactionByName(_FactionName)
-	-- 		for _GuildInitials, _GuildName in pairs (_Guilds) do
-	-- 			XFG:Debug(LogCategory, 'Initializing guild [%s]', _GuildName)
-	-- 			local _NewGuild = Guild:new()
-	-- 			_NewGuild:Initialize()
-	-- 			_NewGuild:SetName(_GuildName)
-	-- 			_NewGuild:SetFaction(_Faction)
-	-- 			_NewGuild:SetRealm(_NewRealm)
-	-- 			_NewGuild:SetInitials(_GuildInitials)
-	-- 			XFG.Guilds:AddGuild(_NewGuild)
-	-- 		end
-	-- 	end
-	-- end
-
-	-- local _RealmName = GetRealmName()
-	-- XFG.Player.Realm = XFG.Realms:GetRealm(_RealmName)
-	-- if(XFG.Player.Realm == nil) then
-	-- 	XFG:Error(LogCategory, 'Player is not on a supported realm [%s]', _RealmName)
-	-- 	return
-	-- end
-
 	-- A significant portion of start up is delayed due to guild information not being available yet
 	XFG.Handlers.TimerEvent = TimerEvent:new(); XFG.Handlers.TimerEvent:Initialize()